+++ src/main/java/com/github/funthomas424242/rades/project/domain/RadesProjectAccessor.java
@@ -17,6 +17,8 @@ public interface RadesProjectAccessor {
 
     String getGithubUsername();
 
+    String getGithubRepositoryname() ;
+
 
 
 }